Add int casts as needed around STACK_OF(T) sizes and indices
We now ensure STACK_OF(T) sizes and indices fit in INT_MAX, so it's safe
to cast to int.

Bug: 516
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/60066
Commit-Queue: David Benjamin <davidben@google.com>

+++ crypto/x509/x_name.c
@@ -230,7 +230,7 @@ static int x509_name_ex_d2i(ASN1_VALUE **val, const unsigned char **in,
     entries = sk_STACK_OF_X509_NAME_ENTRY_value(intname, i);
     for (j = 0; j < sk_X509_NAME_ENTRY_num(entries); j++) {
       entry = sk_X509_NAME_ENTRY_value(entries, j);
-      entry->set = i;
+      entry->set = (int)i;
       if (!sk_X509_NAME_ENTRY_push(nm->entries, entry)) {
         goto err;
       }
